+++ Assets/VRM10_Samples/ClothSample/ClothWarp/Runtime/Jobs/TransformInputOutput.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5f3379ab5e1202d43a95e72ece6bda3a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
